requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.27.2
cryptography==41.0.3
decorator==4.4.2
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
ffmpeg-python==0.2.0
filelock==3.12.2
future==0.18.3
idna==3.4
imageio==2.31.2
imageio-ffmpeg==0.4.8
Jinja2==3.1.2
kombu==5.3.1
lit==16.0.6
llvmlite==0.40.1
MarkupSafe==2.1.3
more-itertools==10.1.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
openai-whisper==20230314
Pillow==10.0.0
proglog==0.1.10
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3
redis==5.0.0
regex==2023.8.8
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tiktoken==0.3.1
torch==2.0.1
torchaudio==2.0.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
